+++ cmd/evm/internal/t8ntool/transition.go
@@ -43,7 +43,7 @@ import (
 	"github.com/ledgerwatch/erigon/common"
 	"github.com/ledgerwatch/erigon/common/math"
 	"github.com/ledgerwatch/erigon/consensus/ethash"
-	"github.com/ledgerwatch/erigon/consensus/serenity"
+	"github.com/ledgerwatch/erigon/consensus/merge"
 	"github.com/ledgerwatch/erigon/core"
 	"github.com/ledgerwatch/erigon/core/state"
 	"github.com/ledgerwatch/erigon/core/types"
@@ -303,9 +303,9 @@ func Main(ctx *cli.Context) error {
 	defer tx.Rollback()
 
 	reader, writer := MakePreState(chainConfig.Rules(0, 0), tx, prestate.Pre)
-	// serenity engine can be used for pre-merge blocks as well, as it
+	// Merge engine can be used for pre-merge blocks as well, as it
 	// redirects to the ethash engine based on the block number
-	engine := serenity.New(&ethash.FakeEthash{})
+	engine := merge.New(&ethash.FakeEthash{})
 
 	result, err := core.ExecuteBlockEphemerally(chainConfig, &vmConfig, getHash, engine, block, reader, writer, nil, getTracer)
